+++ src/main/java/io/github/classgraph/utils/NestedJarHandler.java
@@ -49,7 +49,6 @@ import java.util.zip.ZipFile;
 
 import io.github.classgraph.ModuleReaderProxy;
 import io.github.classgraph.ModuleRef;
-import org.springframework.boot.loader.jar.JarFile;
 
 /**
  * Unzip a jarfile within a jarfile to a temporary file on disk. Also handles the download of jars from http(s) URLs
@@ -798,61 +797,4 @@ public class NestedJarHandler {
         }
     }
 
-    /**
-     * Fake {@link File} for nested jar files handled by Spring-Boot's {@link JarFile}-API.
-     *
-     * This class is only loaded when the Spring-Boot API is available, so it can use the
-     * Spring-Boot types in it's interface.
-     */
-    static class NestedJarFile extends File {
-
-        private final JarFile parentJarFile;
-        private final ZipEntry nestedJarEntry;
-        private final boolean directory;
-
-        NestedJarFile(ZipFile parentJarFile, ZipEntry nestedJarEntry) {
-            this((JarFile) parentJarFile, nestedJarEntry);
-        }
-
-        NestedJarFile(JarFile parentJarFile, ZipEntry nestedJarEntry) {
-            super(parentJarFile.getName() + "!/" + nestedJarEntry.getName());
-            this.parentJarFile = parentJarFile;
-            this.nestedJarEntry = nestedJarEntry;
-            this.directory = nestedJarEntry.isDirectory();
-        }
-
-        JarFile openAsJarFile() throws IOException {
-            return parentJarFile.getNestedJarFile(nestedJarEntry);
-        }
-
-        @Override
-        public boolean canRead() {
-            return true;
-        }
-
-        @Override
-        public boolean canWrite() {
-            return false;
-        }
-
-        @Override
-        public boolean isFile() {
-            return !directory;
-        }
-
-        @Override
-        public boolean isDirectory() {
-            return directory;
-        }
-
-        @Override
-        public boolean exists() {
-            return true;
-        }
-
-        @Override
-        public File getCanonicalFile() {
-            return this;
-        }
-    }
 }
